requirements.txt
# Grocery Pipeline Requirements
# OCR and Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
easyocr>=1.7.0
pytesseract>=0.3.10

# Web Scraping (inherited from parent)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0,<2.2.0

# Data Processing
json5>=0.9.0

# Optional: For better OCR
Pillow>=10.0.0
